+++ docs/CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### What's changed? ❗❗❗
 - You can now choose the placement of the XP Label relative to the XP Bar.
+- If you are at Champion Level, you cannot place XP Label above the XP Bar as it overlays pre-existing text.
+- If XP Label is placed below XP Bar, while hovering the XP Bar the XP Label will hide to show the XP Bar's tooltip. Otherwise, the Label will appear before the tooltip.
 
 ### Features 🚀 
 - Choose where to display the XP Label relative to the XP Bar.
 - Placement Options:
-    - Top
+    - Top ***[Excludes Champion]***
     - Center
     - Bottom